move process model examples to make the tests work
