+++ packages/synthchron-simulator/src/index.ts
@@ -1,4 +1,4 @@
-import { petriNet1 } from '../test/processModelExamples/petriNet';
+import { petriNet1 } from './processModelExamples/petriNet';
 import * as YAML from 'yaml';
 
 export const main = (): string => 'Hello World';
